reload unread channels after ordering
